Fix for Win32, support send() and recv() returning errno=EWOULDBLOCK.

+++ sbbs3/useredit.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet online sysop user editor */
 
-/* $Id: useredit.cpp,v 1.42 2010/03/06 00:13:04 rswindell Exp $ */
+/* $Id: useredit.cpp,v 1.41 2009/11/09 02:54:55 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -682,7 +682,7 @@ int sbbs_t::searchup(char *search,int usernum)
 	if((file=nopen(userdat,O_RDONLY|O_DENYNONE))==-1)
 		return(usernum);
 
-	flen=(long)filelength(file);
+	flen=filelength(file);
 	lseek(file,(long)((long)usernum*U_LEN),0);
 
 	while((i*U_LEN)<=(ulong)flen) {
